# app/controllers/user_controller.go
package controllers

import (
	"cashflow-backend/app/dto"
	"cashflow-backend/app/models"
	"cashflow-backend/pkg/configs"
	"fmt"
	"time"

	"github.com/go-playground/validator/v10"
	"github.com/gofiber/fiber/v2"
	"golang.org/x/crypto/bcrypt"
)

var validate = validator.New()

// Register User
// @Summary Register User
// @Description Endpoint untuk register user
// @Tags User
// @Accept json
// @Produce json
// @Param request body dto.InputRegisterValid true "Input Valid"
// @Success 201 {object} dto.ResponseSuccess  
// @Failure 400 {object} dto.RegisterFailed    
// @Failure 409 {object} dto.RegisterFailed
// @Failure 500 {object} dto.RegisterFailed
func RegisterUser(c *fiber.Ctx) error {
    var input dto.InputRegisterValid
    if err := c.BodyParser(&input); err != nil {
        return c.Status(fiber.StatusBadGateway).JSON(dto.RegisterFailed{
            Message: "Input Tidak Valid",
        })
    }
    
    // ... (Validasi Duplikat & Validator Struct tetap sama) ...
    var existingUser models.User
    if err := configs.DB.Where("email = ? OR username = ?", input.Email, input.Username).First(&existingUser).Error; err == nil {
        return c.Status(fiber.StatusConflict).JSON(dto.RegisterFailed{
            Message: "Email atau Username sudah terdaftar",
        })
    }
    if err := validate.Struct(input); err != nil {
        return c.Status(fiber.StatusBadRequest).JSON(
            // Debugging
            // fiber.Map{"error": "Data tidak lengkap", "detail": err.Error()},
            dto.RegisterFailed{
                Message: "Data tidak lengkap",
            },
        )
    }

    hashedPassword, _ := bcrypt.GenerateFromPassword([]byte(input.Password), bcrypt.DefaultCost)

    // FIX 1: Siapkan variabel role default (UserDefault = 3)
    // Jangan langsung assign angka, harus via variabel biar bisa diambil alamat memorinya (&)
    defaultRole := models.UserDefault
    

    user := models.User{
        Username: input.Username,
        Email:    input.Email,
        Password: string(hashedPassword),
        UserRole: &defaultRole, 
        SubscriptionPlan: models.PlanFree,
    }
    roleVal := models.UserDefault // Default 3
    if user.UserRole != nil {
        roleVal = *user.UserRole // Ambil nilai aslinya (Dereference)
    }

    tx := configs.DB.Begin()
    if err := tx.Create(&user).Error; err != nil {
        tx.Rollback()
        fmt.Println(err)
        return c.Status(fiber.StatusInternalServerError).JSON(
            // Debugging
            // fiber.Map{"message": "Gagal mendaftar", "error" : err.Error()},
            dto.RegisterFailed{
                Message: "Gagal mendaftar",
            },
        )
    }

    defaultWallet := models.Wallet{
        UserID:   user.UserID,
        Name:     "My Wallet",
        Balance:  0,
        Currency: "IDR",
    }
    walletRes := dto.WalletResponse{
        WalletID: defaultWallet.WalletID.String(), // Pastikan convert UUID ke string
        Name:     defaultWallet.Name,
        Balance:  defaultWallet.Balance,
        Currency: defaultWallet.Currency,
    }

    if err := tx.Create(&defaultWallet).Error; err != nil {
        tx.Rollback()
        return c.Status(fiber.StatusInternalServerError).JSON(
            // Debugging
            // fiber.Map{"error": "Gagal membuat wallet default"}
            dto.RegisterFailed{
                Message: "Gagal membuat wallet default",
            },
        )
    }

    tx.Commit()

    // user.Password = ""
    userResponse := dto.UserResponse{
        UserID:           user.UserID.String(), // Convert UUID ke string
        Username:         user.Username,
        Email:            user.Email,
        UserRole:         roleVal,
        RoleText:         "User", // Karena register pasti User biasa
        SubscriptionPlan: user.SubscriptionPlan,
        SubscriptionExp:  user.SubscriptionExp,
        CreatedAt:        user.CreatedAt,
        UpdatedAt:        user.UpdatedAt,
        Wallets:          []dto.WalletResponse{walletRes}, // Masukin wallet tadi
    }
    return c.Status(fiber.StatusCreated).JSON(
        dto.ResponseSuccess{
            Message: "Register Berhasil",
            Data : userResponse,
        },
    )
}

// Get My Profile
// @Summary Get My Profile
// @Description Endpoint untuk mendapatkan profile user
// @Tags User
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 200 {object} dto.UserResponse
// @Failure 400 {object} dto.DefaultMessageResponse
// @Failure 401 {object} dto.DefaultMessageResponse
// @Failure 404 {object} dto.DefaultMessageResponse
// @Failure 500 {object} dto.DefaultMessageResponse
// @Router /users/myprofile [get]
func GetMyProfile(c *fiber.Ctx) error {
    userID, err := getUserIDFromToken(c)
    if err != nil {
        return c.Status(fiber.StatusBadRequest).JSON(dto.DefaultMessageResponse{
            Message : "Token Invalid",
        })
    }

    var user models.User
    result := configs.DB.Preload("Wallets").First(&user, "user_id = ?", userID)
    if result.Error != nil {
        return c.Status(fiber.StatusNotFound).JSON(dto.DefaultMessageResponse{
            Message : "User Not Found",
        },
        )
    }

    // Mapping Wallet Response
    var walletRes []dto.WalletResponse
    for _, w := range user.Wallets {
        walletRes = append(walletRes, dto.WalletResponse{
            WalletID: w.WalletID.String(),
            Name:     w.Name,
            Balance:  w.Balance,
            Currency: w.Currency,
        })
    }

    // FIX 2: Logic Role Text & Pointer Handling
    // roleText := "User"
    // roleValue := models.UserDefault // Default kalau nil

    // if user.UserRole != nil {
    //     roleValue = *user.UserRole // Ambil value dari pointer
        
    //     // Sesuaikan dengan konstanta baru (1=Owner, 2=Admin)
    //     if roleValue == models.UserOwner {
    //         roleText = "Owner"
    //     } else if roleValue == models.UserAdmin {
    //         roleText = "Admin"
    //     }
    // }

    response := dto.UserResponse{
        UserID:           user.UserID.String(),
        Email:            user.Email,
        Username:         user.Username,
        UserRole:         *user.UserRole, // <-- Masukkan value integer (bukan pointer)
        SubscriptionPlan: user.SubscriptionPlan,
        SubscriptionExp:  user.SubscriptionExp,
        CreatedAt:        user.CreatedAt,
        UpdatedAt:        user.UpdatedAt,
        RoleText:         user.UserRole.String(),
        Wallets:          walletRes,
    }
    return c.Status(fiber.StatusOK).JSON(response)
}

// GetAllActiveUser
// @Summary Get All Active User
// @Description Endpoint untuk mendapatkan semua user aktif
// @Tags User
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 200 {object} dto.ResGetUserSuccess
// @Failure 400 {object} dto.DefaultMessageResponse
// @Failure 401 {object} dto.DefaultMessageResponse
// @Failure 404 {object} dto.DefaultMessageResponse
// @Failure 500 {object} dto.DefaultMessageResponse
// @Router /users [get]
func GetAllActiveUser(c *fiber.Ctx) error {
    userId, err := getUserIDFromToken(c)
    if err != nil {
        return c.Status(fiber.StatusBadRequest).JSON(dto.DefaultMessageResponse{
            Message : "Token Invalid",
        })
    }

    var requestor models.User
    if err := configs.DB.First(&requestor, "user_id = ?", userId).Error; err != nil {
        return c.Status(fiber.StatusUnauthorized).JSON(dto.DefaultMessageResponse{
            Message : "Unauthorized",
        })
    }

    // Validasi Role (Admin & Owner boleh akses)
    // Cek nil dulu, baru cek value
    // Logic: Kalau role > Admin (berarti User Biasa/3), tolak.
    if requestor.UserRole == nil || *requestor.UserRole > models.UserAdmin {
        return c.Status(fiber.StatusUnauthorized).JSON(dto.DefaultMessageResponse{
            Message : "Only Admin & Owner can access",
        })
    }
   
    var users []models.User
    if err := configs.DB.Preload("Wallets").Find(&users).Error; err != nil {
        return c.Status(fiber.StatusInternalServerError).JSON(dto.DefaultMessageResponse{
            Message : "Internal Server Error",
        })
    }

    var res []dto.UserResponse
    for _, v := range users {
        
        // FIX 4: Handle pointer role di dalam loop
        rVal := models.UserDefault
        rTxt := "User"
        if v.UserRole != nil {
            rVal = *v.UserRole
            if rVal == models.UserOwner { rTxt = "Owner" }
            if rVal == models.UserAdmin { rTxt = "Admin" }
        }

        var walletRes []dto.WalletResponse
        for _, w := range v.Wallets {
            walletRes = append(walletRes, dto.WalletResponse{
                WalletID: w.WalletID.String(),
                Name:     w.Name,
                Balance:  w.Balance,
                Currency: w.Currency,
            })
        }
        res = append(res, dto.UserResponse{
            UserID:           v.UserID.String(),
            Email:            v.Email,
            Username:         v.Username,
            UserRole:         rVal, // <-- Value
            RoleText:         rTxt, // <-- Text
            SubscriptionPlan: v.SubscriptionPlan,
            SubscriptionExp:  v.SubscriptionExp,
            CreatedAt:        v.CreatedAt,
            UpdatedAt:        v.UpdatedAt,
            Wallets:          walletRes,
        })
    }

    return c.Status(fiber.StatusOK).JSON(dto.ResGetUserSuccess{
        Message: "Fetch Success",
        Data:    res,
    })
}

// Delete User
// @Summary Delete User
// @Description Endpoint untuk Hard Delete User
// @Tags User
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 200 {object} dto.DefaultMessageResponse
// @Failure 400 {object} dto.DefaultMessageResponse
// @Failure 401 {object} dto.DefaultMessageResponse
// @Failure 404 {object} dto.DefaultMessageResponse
// @Failure 500 {object} dto.DefaultMessageResponse
// @Router /users/{id} [delete]
func DeleteUser(c *fiber.Ctx) error {
    reqID, err := getUserIDFromToken(c)
    if err != nil {
        return c.Status(fiber.StatusBadGateway).JSON(dto.DefaultMessageResponse{
            Message : "Token Invalid",
        })
    }
    var requestor *models.User
    if err := configs.DB.First(&requestor, "user_id = ?", reqID).Error; err != nil {
        return c.Status(fiber.StatusUnauthorized).JSON(dto.DefaultMessageResponse{
            Message : "Unauthorized",
        })
    }
    if *requestor.UserRole > models.UserAdmin {
        return c.Status(fiber.StatusUnauthorized).JSON(dto.DefaultMessageResponse{
            Message : "Only Admin can access",
        })
    }

    userId := c.Params("id")
    var user models.User
    // 1. Cek dulu usernya ada atau nggak (Opsional tapi bagus buat UX)
    if err := configs.DB.First(&user, "user_id = ?", userId).Error; err != nil {
        return c.Status(404).JSON(dto.DefaultMessageResponse{
            Message: "User Tidak Ditemukan",
        })
    }

    // 2. Lakukan HARD DELETE dengan Cascade
    // .Unscoped() = Abaikan deleted_at, hapus barisnya dari fisik tabel
    // Karena di Model sudah ada 'OnDelete:CASCADE', database akan otomatis menghapus wallet & transaksi terkait.
    result := configs.DB.Unscoped().Delete(&user)

    if result.Error != nil {
        return c.Status(fiber.StatusInternalServerError).JSON(dto.FailureMessage{
            Message : "Gagal Mengambil Data, Internal Server Error",
            Error: result.Error.Error(),
        
        })
    }

    return c.Status(fiber.StatusOK).JSON(fiber.Map{
        "status": "success",
        "message": "User dan seluruh datanya berhasil dihapus permanen",
    })
}


func UpgradeUserPlan(c *fiber.Ctx) error {
    // 1. Ambil User ID dari Token
    userID, err := getUserIDFromToken(c)
    if err != nil {
        return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{"error": "Unauthorized"})
    }

    // 2. Parse Input Body
    var input dto.UpgradePlanInput
    if err := c.BodyParser(&input); err != nil {
        return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Format input salah"})
    }

    // 3. Validasi Input (Cuma boleh 'premium' atau 'gold')
    if err := validate.Struct(input); err != nil {
        return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
            "error": "Plan tidak valid (pilih: premium, gold)",
            "detail": err.Error(),
        })
    }

    // 4. Cari User di Database
    var user models.User
    if err := configs.DB.First(&user, "user_id = ?", userID).Error; err != nil {
        return c.Status(fiber.StatusNotFound).JSON(fiber.Map{"error": "User tidak ditemukan"})
    }

    // 5. Logic Durasi Berlangganan
    // Misal: Premium = 30 Hari, Gold = 1 Tahun (Contoh saja)
    var duration time.Duration
    if input.Plan == models.PlanGold {
        duration = 365 * 24 * time.Hour // 1 Tahun
    } else {
        duration = 30 * 24 * time.Hour // 30 Hari (Premium)
    }

    // Hitung tanggal kedaluwarsa baru
    // Kalau user masih aktif, tambah dari tanggal exp terakhir. Kalau sudah mati, dari NOW.
    newExpDate := time.Now().Add(duration)
    if user.SubscriptionExp != nil && user.SubscriptionExp.After(time.Now()) {
        // Jika masih aktif, perpanjang dari tanggal exp yang ada
        newExpDate = user.SubscriptionExp.Add(duration)
    }

    // 6. Update Data User
    user.SubscriptionPlan = input.Plan
    user.SubscriptionExp = &newExpDate // Assign pointer

    // Simpan ke Database
    if err := configs.DB.Save(&user).Error; err != nil {
        return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"error": "Gagal mengupdate plan"})
    }

    return c.JSON(fiber.Map{
        "message": "Berhasil upgrade plan",
        "data": fiber.Map{
            "user_id": user.UserID,
            "plan": user.SubscriptionPlan,
            "expires_at": user.SubscriptionExp,
        },
    })
}