# docs/swagger.yaml
basePath: /api
definitions:
  controllers.AddPeopleInput:
    properties:
      email:
        type: string
      group_id:
        type: string
    required:
    - email
    - group_id
    type: object
  controllers.GroupInput:
    properties:
      group_type:
        type: string
      members:
        items:
          type: string
        minItems: 1
        type: array
      name:
        type: string
    required:
    - group_type
    - members
    - name
    type: object
  controllers.GroupResponse:
    properties:
      group_id:
        type: string
      members:
        items:
          $ref: '#/definitions/controllers.MemberResponse'
        type: array
      name:
        type: string
      type:
        type: string
    type: object
  controllers.LoginError:
    properties:
      error:
        example: Record Not Found
        type: string
      message:
        example: Email atau Password Salah
        type: string
    type: object
  controllers.LoginInput:
    properties:
      email:
        example: username@example.com
        type: string
      password:
        example: password
        minLength: 5
        type: string
    required:
    - email
    - password
    type: object
  controllers.LoginSuccess:
    properties:
      message:
        example: Login Success
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3Njk4MTE0NTIsInJvbGUiOjMsInVzZXJfaWQiOiJkN
        type: string
    type: object
  controllers.MemberResponse:
    properties:
      member_id:
        type: string
      role:
        description: Kita pake string (Owner/Member) bukan angka
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  controllers.ServerError:
    properties:
      message:
        example: Server Error
        type: string
    type: object
  dto.DefaultMessageResponse:
    properties:
      message:
        example: Token Invalid
        type: string
    type: object
  dto.InputRegisterValid:
    properties:
      email:
        example: johndoe@example.com
        type: string
      password:
        example: john123
        minLength: 5
        type: string
      username:
        example: johndoe
        minLength: 5
        type: string
    required:
    - email
    - password
    - username
    type: object
  dto.RegisterFailed:
    properties:
      message:
        example: Email atau Username sudah terdaftar
        type: string
    type: object
  dto.ResGetUserSuccess:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
      message:
        example: Berhasil
        type: string
    type: object
  dto.ResponseSuccess:
    properties:
      data:
        $ref: '#/definitions/dto.UserResponse'
      message:
        example: Register Berhasil
        type: string
    type: object
  dto.UserResponse:
    properties:
      created_at:
        example: "2023-06-30T00:00:00Z"
        type: string
      email:
        example: johndoe@example.com
        type: string
      role_text:
        example: User
        type: string
      subscription_exp:
        example: "2023-06-30T00:00:00Z"
        type: string
      subscription_plan:
        example: free
        type: string
      updated_at:
        example: "2023-06-30T00:00:00Z"
        type: string
      user_id:
        example: a0b1c2d3-e4f5-g6h7-i8j9-k0l1m2n3
        type: string
      user_role:
        allOf:
        - $ref: '#/definitions/models.UserRole'
        example: 3
      username:
        example: JohnDoe
        type: string
      wallet:
        items:
          $ref: '#/definitions/dto.WalletResponse'
        type: array
    type: object
  dto.WalletResponse:
    properties:
      balance:
        example: 100000
        type: number
      currency:
        example: IDR
        type: string
      wallet_id:
        example: a0b1c2d3-e4f5-g6h7-i8j9-k0l1m2n3
        type: string
      wallet_name:
        example: wallet1
        type: string
    type: object
  models.UserRole:
    enum:
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-comments:
      UserAdmin: Bisa atur user
      UserDefault: Pengguna aplikasi
      UserOwner: Dewa
    x-enum-descriptions:
    - Dewa
    - Bisa atur user
    - Pengguna aplikasi
    x-enum-varnames:
    - UserOwner
    - UserAdmin
    - UserDefault
host: localhost:3000
info:
  contact:
    email: bagasramadhan239@gmail.com
    name: Bagas Ramadhan Rusnadi
  description: Dokumentasi API untuk aplikasi Cashflow.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Cashflow API Documentation
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Endpoint untuk login
      parameters:
      - description: Login Input
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.LoginSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.LoginError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.LoginError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ServerError'
      summary: Login
      tags:
      - Auth
  /groups:
    post:
      consumes:
      - application/json
      description: Endpoint untuk User membuat grup baru dan otomatis menjadi group-owner.
      parameters:
      - description: Data Group (Minimal 1 member)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.GroupInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.GroupResponse'
        "400":
          description: Error validasi / ID member salah
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Token tidak valid
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Membuat Group Baru
      tags:
      - Groups
  /groups/add-new-member:
    post:
      consumes:
      - application/json
      description: Menambahkan anggota baru ke Group berdasarkan email
      parameters:
      - description: AddPeopleInput
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.AddPeopleInput'
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil menambahkan anggota
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Input Tidak Valid
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized / Invalid Token
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Anda bukan anggota grup ini
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Email user tidak ditemukan di sistem.
          schema:
            additionalProperties: true
            type: object
        "409":
          description: User ini sudah menjadi anggota grup.
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Gagal menambahkan anggota baru.
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Menambahkan anggota baru ke Group
      tags:
      - Groups
  /users:
    get:
      consumes:
      - application/json
      description: Endpoint untuk mendapatkan semua user aktif
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResGetUserSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.DefaultMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.DefaultMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.DefaultMessageResponse'
      security:
      - BearerAuth: []
      summary: Get All Active User
      tags:
      - User
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Endpoint untuk Hard Delete User
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DefaultMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.DefaultMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.DefaultMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.DefaultMessageResponse'
      security:
      - BearerAuth: []
      summary: Delete User
      tags:
      - User
  /users/myprofile:
    get:
      consumes:
      - application/json
      description: Endpoint untuk mendapatkan profile user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.DefaultMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.DefaultMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.DefaultMessageResponse'
      security:
      - BearerAuth: []
      summary: Get My Profile
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: "Bearer <token_jwt_disini>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"